Adding a data-quality monitor to check the status of a switch. WIP.
When a switch connected to CTS is pressed the a file is created, signalling to the
AuroraWatchNet daemon that data quality may be impaired for some reason. Pressing
the switch again toggles state and removes the file. This python script handles the
daemon start/stop/restart and the actual work.

Work in progress. Need to drop root priviledges.