Adding DHCP, DNS lookups and more.
Initial ethernet configuration uses DHCP if the local IP address
defined in the EEPROM is 0.0.0.0.

Manual configuration now requires local IP, netmask, gateway and up to
3 DNS servers. If the local IP address is set to 0.0.0.0 in the EEPROM
then DHCP is used. The single DNS server IP that is obtained from the
Ethernet class replaces the primary DNS server defined in the EEPROM
settings. The secondary and tertiary DNS servers are kept.

If the remote (server) hostname is defined it will be looked up using
DNS. If necessary multiple attempts using up to 3 different DNS
servers will be made. If the remote hostname is not set, or it could
not be resolved, then the remote IP setting from the EEPROM will be
used. If the remote IP setting is zero a reboot will be forced until
the hostname can be resolved.

The microcontroller will not sleep if DHCP is in operation; the DHCP
lease maintenance uses millis(), and so is not compatibile with sleep
mode.