Call wdt_reset if ETHERNET2_USE_WDT is defined
The maximum timeout that can be set for the watchdog timer is 8
seconds. On a busy or slow network it is possible that DHCP requests
or renewals take longer than this, thereby causing an unwanted reboot
(or worse, continual reboots). Disabling the WDT before DHCP lease
requests/renewals is possible but undesirable.