Split manual into Calunium and Raspberry Pi versions
Split the manual into two versions, one where the sensor head is
connected to a Calunium microcontroller and one where the sensor
head is connected directly to the Raspberry Pi. Note that the Calunium
version also makes use of a Raspberry Pi for the base unit, but the
sensor head is not connected directly to the Pi.

The two versions are generated from the same source files; different
job names are assigned which then select the appropriate conditional
compilation units. The Verbatim environments cannot be used inside
the conditional compilation units but the VerbatimInput command (and
derivatives defined by CustomVerbatimCommand} can be.

Work in progress: only the Raspberry Pi setup chapter currently reflects
the differences.