[doc] Document how to configure private network
When a separate microcontroller is used for data acquisition it must be
programmed with the IP address or hostname of the Raspberry Pi data
logger. This can be a problem when the full network details are not
known before deployment. In such cases it is helpful to install a
private network between the microcontroller and Raspberry Pi.